ch4/ofi: refactor ofi/init_addrxchg.c
The address exchange and av insertion code become more complex due to
the additions of multi-nic and multi-vni. However, it is fairly
abstracted from the rest of the init process. Move the code to separate
unit to ease the maintenance.
