+++ src/mpid/ch4/netmod/ofi/ofi_init.h
@@ -32,4 +32,7 @@ void MPIDI_OFI_update_global_settings(struct fi_info *prov);
 /* Determine if NIC has already been included in others */
 bool MPIDI_OFI_nic_already_used(const struct fi_info *prov, struct fi_info **others, int nic_count);
 
+int MPIDI_OFI_addr_exchange_root_ctx(MPIR_Comm * init_comm);
+int MPIDI_OFI_addr_exchange_all_ctx(void);
+
 #endif /* OFI_INIT_H_INCLUDED */
